+++ api/db.json
@@ -1,19 +1,95 @@
 {
   "expenseTypes": [
     {
-      "name": "Food Edits",
+      "name": "Grocery",
       "need": "Primary",
       "id": 1
     },
     {
       "name": "Shelter",
       "need": "Primary",
+      "id": 2
+    },
+    {
+      "name": "Car",
+      "need": "Primary",
       "id": 3
     },
     {
-      "name": "Ads",
+      "name": "Clothing",
       "need": "Primary",
       "id": 4
+    },
+    {
+      "name": "Entertainment",
+      "need": "Secondary",
+      "id": 5
+    },
+    {
+      "name": "Amazon And Other Online Shopping",
+      "need": "Misc",
+      "id": 6
+    },
+    {
+      "name": "Adds New",
+      "need": "Secondary",
+      "id": 7
+    }
+  ],
+  "expenses": [
+    {
+      "id": 1,
+      "date": "09/15/2019",
+      "description": "Rent",
+      "type": "Shelter",
+      "mode": "Fixed",
+      "amount": 1000,
+      "notes": "recurring one"
+    },
+    {
+      "id": 2,
+      "date": "09/6/2019",
+      "description": "Gas",
+      "type": "Car",
+      "mode": "Variable",
+      "amount": 100,
+      "notes": "non recurring one"
+    },
+    {
+      "date": "09/15/2019",
+      "description": "walmart",
+      "type": "Grocery",
+      "mode": "variable",
+      "amount": 100,
+      "notes": "recurring one",
+      "id": 7
+    },
+    {
+      "date": "09/16/2019",
+      "description": "fresh",
+      "type": "Shelter",
+      "mode": "Fixed",
+      "amount": "12",
+      "notes": "fresh expense ",
+      "id": 8
+    },
+    {
+      "date": "09/16/2019",
+      "description": "new",
+      "type": "Grocery",
+      "mode": "Fixed",
+      "amount": "12",
+      "notes": "12",
+      "id": 9
+    },
+    {
+      "date": "09/16/2019",
+      "description": "qwerty",
+      "type": "Clothing",
+      "mode": "Fixed",
+      "amount": "1212",
+      "notes": "qwerty",
+      "id": 10
     }
   ]
 }
