+++ src/components/common/radio.jsx
@@ -10,14 +10,13 @@ const Radio = ({
 	...rest
 }) => {
 	return (
-		<div className="col-auto">
+		<div onChange={onChange} className="col-auto">
 			<label htmlFor={labelFor}>{labelName}</label>
 			{options.map(option => (
 				<div>
 					<input
 						type={type}
-						id={inputId}
-						onChange={onChange}
+						id={option}
 						name={inputId}
 						value={option}
 						{...rest}
