+++ src/reducers/expense.js
@@ -4,33 +4,40 @@ import {
     DELETE_EXPENSE,
     UPDATE_EXPENSE
 } from "../constants/constants";
+import _ from "lodash";
 
-const initialState = [{
-    id: 1,
-    date: "09/ 15 / 2019",
-    description: "rent",
-    type: "primary",
-    mode: "fixed",
-    amount: 1000,
-    notes: "recurring one"
-}];
+const initialState = {
+    expenses: []
+}
 
 const expense = (state = initialState, action) => {
     switch (action.type) {
-        case "GET_EXPENSE":
-            return null;
+        case GET_EXPENSE:
 
-        case "ADD_EXPENSE":
-            return null
+            if (_.isEqual(action.payload.sort(), state.expenses.sort())) {
+                return Object.assign({}, state, {
+                    expenses: state.expenses
+                });
+            } else {
+                return Object.assign({}, state, {
+                    expenses: state.expenses.concat(action.payload)
+                });
+            }
+            case ADD_EXPENSE:
+                return Object.assign({}, state, {
+                    expenses: state.expenses.concat({
+                        ...action.payload
+                    })
+                });
 
-        case "UPDATE_EXPENSE":
-            return null
+            case UPDATE_EXPENSE:
+                return null
 
-        case "DELETE_EXPENSE":
-            return null
+            case DELETE_EXPENSE:
+                return null
 
-        default:
-            return state;
+            default:
+                return state;
     }
 
 }
