+++ src/services/expenseService.js
@@ -1,0 +1,10 @@
+import http from "../httpAxiosModule/httpAxiosModule";
+import config from "../config.json";
+
+export const getExpenseService = () => {
+    return http.get(config.apiEndPoint + "/expenses");
+}
+
+export const addExpenseService = (expense) => {
+    return http.post(config.apiEndPoint + "/expenses", expense);
+}
